Extract GraphQL setup to separate module with minimal schema
- Create backend/graphql.ts with basic GraphQL Yoga setup
- Remove mutations, hello query, notebooks query - keep only me query
- Clean separation of concerns for future GraphQL expansion
- Maintain auth context and error handling
- Update selective-entry.ts to import from new module

Note: Tests need type fixing for Request/Response compatibility
